Add Import raw to parquet for os open uprn, added test, adjusted type in clean case class

+++ data-source-os-open-uprn/os-open-uprn-model/src/main/scala/com/landalytics/model/osopenuprn/clean/CleanOsOpenUprnModel.scala
@@ -4,8 +4,8 @@ object CleanOsOpenUprnModel {
 
   case class OsOpenUprn(
                          uprn: String,
-                         xCoordinate: Int,
-                         yCoordinate: Int,
+                         xCoordinate: Double,
+                         yCoordinate: Double,
                          latitude: Double,
                          longitude: Double
                        )
